+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/UtgiftBeregning.kt
@@ -1,5 +1,6 @@
 package no.nav.tilleggsstonader.sak.vedtak
 
+import no.nav.tilleggsstonader.kontrakter.felles.Datoperiode
 import no.nav.tilleggsstonader.kontrakter.felles.Periode
 import java.time.LocalDate
 import java.time.YearMonth
@@ -9,7 +10,9 @@ sealed class UtgiftBeregningType<T>(
     override val fom: T,
     override val tom: T,
     open val utgift: Int,
-) : Periode<T> where T : Comparable<T>, T : Temporal
+) : Periode<T> where T : Comparable<T>, T : Temporal {
+    abstract fun tilDatoPeriode(): Datoperiode
+}
 
 data class UtgiftBeregningMåned(
     override val fom: YearMonth,
@@ -19,6 +22,8 @@ data class UtgiftBeregningMåned(
     init {
         validatePeriode()
     }
+
+    override fun tilDatoPeriode(): Datoperiode = Datoperiode(fom = fom.atDay(1), tom = tom.atEndOfMonth())
 }
 
 data class UtgiftBeregningDato(
@@ -29,4 +34,6 @@ data class UtgiftBeregningDato(
     init {
         validatePeriode()
     }
+
+    override fun tilDatoPeriode(): Datoperiode = Datoperiode(fom = fom, tom = tom)
 }
