+++ src/test/resources/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/beregning_utgifter.feature
@@ -143,4 +143,4 @@ Egenskap: Beregning - Utgifter
 
     Når beregner
 
-    Så forvent følgende feil: Kan ikke innvilge når det ikke finnes utgifter hele vedtaksperioden
+    Så forvent følgende feil: Vedtaksperioden 01.01.2023–28.02.2023 mangler oppfylt utgift hele eller deler av perioden.
